+++ CHANGELOG.md
@@ -6,6 +6,11 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Changed
+
+- Default X/Twitter proxy is now `vxtwitter.com`. `fxtwitter.com` was serving
+  degraded embeds without video metadata, so videos did not play inline.
+
 ## [0.1.0] - 2026-06-01
 
 ### Added